// frontend_business/src/core/enums/weddingVendorType.enum.ts
export enum WeddingVendorTypeEnum {
	NONE = "NONE",
	Photographer = "Photographer", // Фотограф, который будет снимать свадьбу
	Videographer = "Videographer", // Видеограф, который будет снимать видео
	Caterer = "Caterer", // Кейтеринг, который будет обеспечивать еду и напитки
	Florist = "Florist", // Флорист, который будет заниматься цветочными композициями
	DJ = "DJ", // Диджей, который будет отвечать за музыку
	Band = "Band", // Музыкальная группа, если вы предпочитаете живую музыку
	WeddingPlanner = "WeddingPlanner", // Организатор свадьбы, который поможет с планированием и координацией
	Venue = "Venue", // Место проведения свадьбы
	Baker = "Baker", // Пекарь, который будет делать свадебный торт
	MakeupArtist = "MakeupArtist", // Визажист, который будет заниматься макияжем
	HairStylist = "HairStylist", // Парикмахер, который будет заниматься прической
	Officiant = "Officiant", // Человек, который будет проводить церемонию
	Transportation = "Transportation", // Транспортные услуги для гостей и молодоженов
	Decorator = "Decorator", // Декоратор, который будет заниматься оформлением зала
	BridalShop = "BridalShop", // Свадебный магазин, где можно приобрести свадебные платья и аксессуары
}